Migrate from urxvt to kitty
Since urxvt is a more or less unmaintained mess, it's difficult to get
cool features to work. For example, it never supported emojis. Or
there's at least five patched versions to install.
On the other hand, kitty is a very new terminal emulator, but it already
has several advanced features like built-in image previews, a diff mode
and - most importantly - support for custom hotkeys.
So, let's remove everything urxvt and introduce kitty as the default
terminal emulator. 😁